requirements.txt

Django
ipython
requests
django-environ
gunicorn
django-celery-results
redis
django-celery-beat
django-redis

